requirements.txt
backports.entry-points-selectable==1.1.1
bidict==0.21.4
blis==0.7.5
catalogue==2.0.6
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.9
click==7.1.2
colorama==0.4.4
contextvars==2.4
cymem==2.0.6
dataclasses==0.8
distlib==0.3.4
dnspython==2.1.0
eventlet==0.25.1
filelock==3.4.0
Flask==1.1.1
Flask-Cors==3.0.10
Flask-SocketIO==4.2.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
idna==3.3
immutables==0.16
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
langcodes==3.3.0
MarkupSafe==2.0.1
mathparse==0.1.2
monotonic==1.6
murmurhash==1.0.6
numpy==1.19.5
packaging==21.3
pathy==0.6.1
platformdirs==2.4.0
preshed==3.0.6
pycparser==2.21
pydantic==1.8.2
pyparsing==3.0.6
python-dateutil==2.8.2
python-engineio==4.3.0
python-socketio==5.5.0
pytz==2021.3
PyYAML==3.13
requests==2.26.0
simple-websocket==0.5.0
six==1.16.0
smart-open==5.2.1
spacy==3.0.7
spacy-legacy==3.0.8
spacy-loggers==1.0.1
SQLAlchemy==1.3.24
srsly==2.4.2
thinc==8.0.13
tqdm==4.62.3
typer==0.3.2
typing-extensions==3.10.0.2
urllib3==1.26.7
virtualenv==20.10.0
wasabi==0.9.0
websocket-client==1.2.3
Werkzeug==2.0.2
wincertstore==0.2
wsproto==1.0.0
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
